refactor(send): extract LNURL input detection + resolution to fit cap
SendSheet is over the 1000-line cap; #793's LNURL additions pushed it to 1264
(baseline 1176), failing the file-size gate. Extract the cohesive pieces:
- src/utils/sendSheetInput.ts — the pure input detectors/decoders
  (DecodedInvoice, decodeInvoice, isLightningAddress, isValidInvoice,
  isLnurlString) + a unit test.
- src/hooks/useSendSheetLnurl.ts — the lightning-address + raw-LNURL
  resolution effects (moved verbatim; behaviour unchanged).
SendSheet 1264 -> 1143 (under baseline). No behaviour change.
